% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoronoiVerse.R
\name{VoronoiVerse}
\alias{VoronoiVerse}
\title{Generates a Voronoi diagram with random seed points.}
\usage{
VoronoiVerse(
  numPoints = 100,
  numSeeds = 20,
  minX = 0,
  maxX = 10,
  minY = 0,
  maxY = 10,
  palette = rainbow(numSeeds),
  ...
)
}
\arguments{
\item{numPoints}{The number of points to generate in the grid.}

\item{numSeeds}{The number of random seed points to generate.}

\item{minX}{The minimum x-coordinate value for the grid.}

\item{maxX}{The maximum x-coordinate value for the grid.}

\item{minY}{The minimum y-coordinate value for the grid.}

\item{maxY}{The maximum y-coordinate value for the grid.}

\item{palette}{The color palette for the Voronoi diagram.
Default is rainbow(numSeeds).}

\item{...}{Additional arguments to be passed to the image function.}
}
\value{
None (plot is displayed).
}
\description{
Generates a Voronoi diagram with random seed points.
}
\examples{
# Generate a Voronoi diagram with defalt parameters
VoronoiVerse(100, 20, 0, 10, 0, 10)

# Generate a Voronoi diagram with a custom color palette
VoronoiVerse(100, 20, 0, 10, 0, 10, palette = c("#AAE6AF", "#AAE6D9", "#AAADE6", "#E6AAE6", "#E6AAAA"))

}
